.config/.config/script/mdtranslate/requirements.txt
python-sentence-splitter
python-openai
python-tiktoken
python-httpx==0.27.2
# git+https://github.com/ssut/py-googletrans.git
